Add files via upload
This notebook contains all the code needed to reproduce the figures in our chapter of the Encyclopaedia of Astrophysics chapter "The satellite galaxies of the Milky Way and Andromeda". It relies on the local_volume_database (https://github.com/apace7/local_volume_database) maintained by Andrew Pace. We endeavour to alter our code if this repository changes.